Add page feature files, rename useInputValid hook
