+++ src/CreateStudyComponents/CreateFooter.js
