+++ src/CreateStudyComponents/CreateForm.js
@@ -6,7 +6,7 @@ import CreatePassword from './CreatePassword.js';
 import CreateFooter from './CreateFooter.js';
 import backgroundGreen from '../img/background_1.png';
 
-import useInputValid from '../hooks/SignupValidation.js';
+import useInputValid from '../hooks/useInputValid.js';
 
 function CreateForm() {
   const [values, setValues] = useState({
